refactor: 패키지 이동 및 시간 형식 변경 클래스 별도로 분리
